refactor: 💡 back to home if repositories is null

+++ src/pages/Repositories/index.js
@@ -1,17 +1,22 @@
 import React, { useEffect, useState } from "react";
-import { Link } from "react-router-dom";
+import { Link, useHistory } from "react-router-dom";
 
 import "./repositories.css";
 
 function Repositories() {
+  const history = useHistory();
   const [repositories, setRepositories] = useState([]);
 
   useEffect(() => {
     let repositoriesName = localStorage.getItem("repositoriesName");
-    repositoriesName = JSON.parse(repositoriesName);
+    if (repositoriesName !== null) {
+      repositoriesName = JSON.parse(repositoriesName);
 
-    setRepositories(repositoriesName);
-    localStorage.clear();
+      setRepositories(repositoriesName);
+      localStorage.clear();
+    } else {
+      history.push("/");
+    }
   }, []);
 
   return (
